pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "BoGuan"
version = "1.0.0"
description = "BoGuan - 基于 DeepAgents 框架的智能运维 Agent 管理与对话系统"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "kanlishiyi"},
]
keywords = ["ai-agent", "deepagents", "mcp", "langchain", "monitoring", "ops", "BoGuan"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "deepagents>=0.4.11",
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-mcp-adapters>=0.2.0",
    "langgraph>=1.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.30.0",
    "sse-starlette>=3.0.0",
    "pydantic>=2.0.0",
    "fpdf2>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/kanlishiyi/BoGuan"
Repository = "https://github.com/kanlishiyi/BoGuan"
Issues = "https://github.com/kanlishiyi/BoGuan/issues"

[project.scripts]
boguan-platform = "boguan.app:main"

[tool.setuptools.packages.find]
include = ["boguan*"]

[tool.setuptools.package-data]
boguan = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
